+++ src/prefect/engine/task_runner.py
@@ -2,6 +2,8 @@
 
 import collections
 import pendulum
+import threading
+from functools import partial, wraps
 from typing import Any, Callable, Dict, Iterable, List, Union, Set, Sized, Optional
 
 import prefect
@@ -17,6 +19,7 @@ from prefect.engine.state import (
     Mapped,
     Pending,
     Scheduled,
+    Resume,
     Retrying,
     Running,
     Skipped,
@@ -26,7 +29,7 @@ from prefect.engine.state import (
     TriggerFailed,
 )
 from prefect.engine.runner import ENDRUN, Runner, call_state_handlers
-from prefect.utilities.executors import main_thread_timeout
+from prefect.utilities.executors import run_with_heartbeat, main_thread_timeout
 
 
 class TaskRunner(Runner):
@@ -68,6 +71,10 @@ class TaskRunner(Runner):
         self.client = Client()
         self.result_handler = result_handler
         super().__init__(state_handlers=state_handlers)
+
+    def _heartbeat(self) -> None:
+        task_run_id = prefect.context.get("_task_run_id")
+        self.client.update_task_run_heartbeat(task_run_id)
 
     def call_runner_target_handlers(self, old_state: State, new_state: State) -> State:
         """
@@ -205,8 +212,9 @@ class TaskRunner(Runner):
         with prefect.context(context, _task_name=self.task.name, _map_index=map_index):
 
             try:
-                # retrieve the run number and place in context
-                state = self.get_run_count(state=state)
+                # retrieve the run number and place in context,
+                # or put resume in context if needed
+                state = self.update_context_from_state(state=state)
 
                 # check if all upstream tasks have finished
                 state = self.check_upstream_finished(
@@ -292,11 +300,15 @@ class TaskRunner(Runner):
         return state
 
     @call_state_handlers
-    def get_run_count(self, state: State) -> State:
+    def update_context_from_state(self, state: State) -> State:
         """
+        Updates context with information contained in the task state:
+
         If the task is being retried, then we retrieve the run count from the initial Retry
         state. Otherwise, we assume the run count is 1. The run count is stored in context as
         _task_run_count.
+
+        Also, if the task is being resumed through a `Resume` state, updates context to have `resume=True`.
 
         Args:
             - state (State): the current state of the task
@@ -308,6 +320,8 @@ class TaskRunner(Runner):
             run_count = state.run_count + 1
         else:
             run_count = 1
+        if isinstance(state, Resume):
+            prefect.context.update(resume=True)
         prefect.context.update(_task_run_count=run_count)
         return state
 
@@ -615,6 +629,7 @@ class TaskRunner(Runner):
 
         return Running(message="Starting task run.")
 
+    @run_with_heartbeat
     @call_state_handlers
     def get_task_run_state(
         self, state: State, inputs: Dict[str, Any], timeout_handler: Optional[Callable]